requirements.txt
sympy==1.7.1
azure-identity
azure-keyvault-secrets